refactor: extract SubsonicPodcastController — fixes #97
